+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/messages/lp/components/PoolLPCard.tsx
@@ -1,9 +1,8 @@
 import { Flex } from "@chakra-ui/react";
 
-import { PoolAssetCard } from "../../components";
+import { ErrorFetchingDetail, PoolAssetCard } from "../../components";
 import { getPoolDenom } from "../../utils";
 import { Loading } from "lib/components/Loading";
-import { EmptyState } from "lib/components/state";
 import { useTxData } from "lib/services/txService";
 import type { AssetInfos, Option } from "lib/types";
 import { coinToTokenWithValue } from "lib/utils";
@@ -30,21 +29,26 @@ export const PoolLPCard = ({
   ampCopierSection,
 }: PoolLPCardProps) => {
   const { data: txData, isLoading } = useTxData(txHash, isOpened);
-  if (!msgShareAmount && isLoading) return <Loading withBorder={false} />;
+  if (txHash && isLoading) return <Loading withBorder={false} />;
 
   const poolDenom = getPoolDenom(poolId);
-  const eventShareAmount = txData?.logs
-    .find((log) => log.msg_index === msgIndex)
-    ?.events?.find((event) => event.type === "coin_received")
-    ?.attributes[3]?.value.slice(0, -poolDenom.length);
+
+  const msgEvents = txData?.logs.find(
+    (log) => log.msg_index === msgIndex
+  )?.events;
+  const eventShareAmount = msgEvents
+    ?.find(
+      (event) =>
+        event.type === "coin_received" &&
+        event.attributes.some((attr) => attr.value.includes(poolDenom))
+    )
+    ?.attributes.find((attr) => attr.key === "amount")
+    ?.value.slice(0, -poolDenom.length);
 
   const shareAmount = msgShareAmount ?? eventShareAmount;
-  if (!shareAmount)
-    return (
-      <EmptyState message="There is an error during fetching transaction detail." />
-    );
-  const poolToken = coinToTokenWithValue(poolDenom, shareAmount, assetInfos);
+  if (!shareAmount) return <ErrorFetchingDetail />;
 
+  const poolToken = coinToTokenWithValue(poolDenom, shareAmount, assetInfos);
   return (
     <Flex
       className="pool-msg-detail-container"