+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/messages/swap/components/PoolSwap.tsx
@@ -1,23 +1,58 @@
 import { Grid, Text } from "@chakra-ui/react";
+import type { Event } from "@cosmjs/stargate";
+import type { Coin } from "cosmjs-types/cosmos/base/v1beta1/coin";
 
-import { AssetCard } from "../../components";
+import { AssetCard, ErrorFetchingDetail } from "../../components";
 import { CustomIcon } from "lib/components/icon";
 import { Loading } from "lib/components/Loading";
-import { EmptyState } from "lib/components/state";
 import { useTxData } from "lib/services/txService";
 import type { AssetInfos, Option } from "lib/types";
 import { coinsFromStr } from "lib/utils";
 
+interface ExactInput {
+  isExactIn: boolean;
+  amount: Coin;
+  expectedDenom: string;
+}
+
 interface PoolSwapInterface {
   txHash: string;
+  exactInput: ExactInput;
   msgIndex: number;
   assetInfos: Option<AssetInfos>;
   isOpened: boolean;
   ampCopierSection?: string;
 }
 
+const getAssets = (
+  exactInput: ExactInput,
+  msgEvents: Option<readonly Event[]>
+): { inAsset: Coin; outAsset: Coin } => {
+  if (exactInput.isExactIn) {
+    return {
+      inAsset: exactInput.amount,
+      outAsset: coinsFromStr(
+        msgEvents
+          ?.findLast((event) => event.type === "coin_received")
+          ?.attributes.find((attr) => attr.key === "amount")?.value ??
+          `0${exactInput.expectedDenom}`
+      )[0],
+    };
+  }
+  return {
+    inAsset: coinsFromStr(
+      msgEvents
+        ?.find((event) => event.type === "coin_spent")
+        ?.attributes.find((attr) => attr.key === "amount")?.value ??
+        `0${exactInput.expectedDenom}`
+    )[0],
+    outAsset: exactInput.amount,
+  };
+};
+
 export const PoolSwap = ({
   txHash,
+  exactInput,
   msgIndex,
   assetInfos,
   isOpened,
@@ -25,24 +60,13 @@ export const PoolSwap = ({
 }: PoolSwapInterface) => {
   const { data: txData, isLoading } = useTxData(txHash, isOpened);
   if (isLoading) return <Loading withBorder={false} />;
+  if (!txData) return <ErrorFetchingDetail />;
 
-  const swapEvent = txData?.logs
-    .find((log) => log.msg_index === msgIndex)
-    ?.events?.find((event) => event.type === "token_swapped");
-  if (!swapEvent)
-    return (
-      <EmptyState message="There is an error during fetching message detail." />
-    );
+  const msgEvents = txData.logs.find(
+    (log) => log.msg_index === msgIndex
+  )?.events;
 
-  // Get the token-in from the third attribute of the event e.g. 10000utoken
-  const inAsset = swapEvent.attributes[3]?.value ?? "";
-  const { amount: inAmount, denom: inDenom } = coinsFromStr(inAsset)[0];
-
-  // Get the token-out from the last attribute of the event e.g. 10000utoken
-  const outAsset =
-    swapEvent.attributes[swapEvent.attributes.length - 1]?.value ?? "";
-  const { amount: outAmount, denom: outDenom } = coinsFromStr(outAsset)[0];
-
+  const { inAsset, outAsset } = getAssets(exactInput, msgEvents);
   return (
     <Grid
       gap={4}
@@ -55,9 +79,9 @@ export const PoolSwap = ({
           From
         </Text>
         <AssetCard
-          amount={inAmount}
-          denom={inDenom}
-          assetInfo={assetInfos?.[inDenom]}
+          amount={inAsset.amount}
+          denom={inAsset.denom}
+          assetInfo={assetInfos?.[inAsset.denom]}
           ampCopierSection={ampCopierSection}
         />
       </div>
@@ -67,9 +91,9 @@ export const PoolSwap = ({
           To
         </Text>
         <AssetCard
-          amount={outAmount}
-          denom={outDenom}
-          assetInfo={assetInfos?.[outDenom]}
+          amount={outAsset.amount}
+          denom={outAsset.denom}
+          assetInfo={assetInfos?.[outAsset.denom]}
           ampCopierSection={ampCopierSection}
         />
       </div>